backend/requirements.txt
fastapi
uvicorn
pymongo 
markdown2
motor
python-dotenv
python-multipart